requirements.txt
langchain 
langchain-community 
langchain-groq
chromadb  
pypdf 
sentence-transformers  
python-dotenv
streamlit